% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_utils.R
\name{af_add_brackets_to_levels}
\alias{af_add_brackets_to_levels}
\title{Add Brackets to Factor Levels}
\usage{
af_add_brackets_to_levels(x)
}
\arguments{
\item{x}{A vector, potentially a factor}
}
\value{
A factor with levels enclosed in square brackets, or the original input if not a factor
}
\description{
This function adds square brackets around the levels of a factor variable.
If the input is not a factor, it returns the input unchanged.
}
\examples{
# Example with a factor
f <- factor(c("A", "B", "C"))
af_add_brackets_to_levels(f)
# Example with a non-factor
v <- c("A", "B", "C")
af_add_brackets_to_levels(v)

}
