% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_factor_analysis.R
\name{af_cfa_between_group}
\alias{af_cfa_between_group}
\title{Perform CFA for Multiple Groups}
\usage{
af_cfa_between_group(df, group_var, latent_var, cfa_vars, model = NULL)
}
\arguments{
\item{df}{Data frame containing the data for CFA}

\item{group_var}{Name of the grouping variable}

\item{latent_var}{Name of the latent variable}

\item{cfa_vars}{List of manifest variable names}

\item{model}{(Optional) Custom CFA model specification}
}
\value{
List containing updated data frame with latent scores, fit results by group,
CFA results table, distribution plots, and correlation plots
}
\description{
Performs confirmatory factor analysis (CFA) separately for multiple groups
within a dataset, generating fit measures, predicted latent scores,
and visualizations for each group.
}
