% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_factor_analysis.R
\name{af_check_order}
\alias{af_check_order}
\title{Sanity check internal function}
\usage{
af_check_order(
  df,
  fit_predict,
  group,
  latent_var_name,
  var_list,
  latent_scores
)
}
\arguments{
\item{df}{Original dataframe}

\item{fit_predict}{List of predicted values from lavaan::lavPredict}

\item{group}{Grouping variable name}

\item{latent_var_name}{Name of the latent variable}

\item{var_list}{List of manifest variable names}

\item{latent_scores}{Vector of latent scores}
}
\value{
List indicating if the order is correct and any errors found
}
\description{
Checks that the ordering of predicted values from lavaan::lavPredict
matches the original dataframe ordering.
}
