% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_basic_plots.R
\name{af_create_custom_legend}
\alias{af_create_custom_legend}
\title{Create custom legend grob}
\usage{
af_create_custom_legend(legend_specs, labels)
}
\arguments{
\item{legend_specs}{List of color-linetype pairs}

\item{labels}{Character vector of labels}
}
\value{
A grob object representing the legend
}
\description{
Create a custom legend from color and line type specifications
}
\examples{
#' legend_specs <- list(
 c("red", "dashed"),
c("blue", "dotted"),
c("green", "solid")
)
labels <- c("Treatment A", "Treatment B", "Control")
legend_grob <- af_create_custom_legend(legend_specs, labels)
grid.newpage()
grid.draw(legend_grob)

}
