% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_descriptive.R
\name{af_create_data_dictionary}
\alias{af_create_data_dictionary}
\title{Create Data Dictionary}
\usage{
af_create_data_dictionary(df)
}
\arguments{
\item{df}{(data.frame) A data frame to create a data dictionary for}
}
\value{
(gt table) A gt table summarizing the data dictionary information
}
\description{
Generates a data dictionary for a given data frame, summarizing variable names,
labels, types, and ranges or levels. The resulting information is formatted
as a gt table for easy viewing.
}
\examples{
data_dict <- af_create_data_dictionary(mtcars)
data_dict

}
