% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_packages.R
\name{af_create_package}
\alias{af_create_package}
\title{Automate Initial Package Creation}
\usage{
af_create_package(
  package_name,
  package_path = ".",
  github_repo = NULL,
  author_name = "Your Name",
  author_email = "you@example.com",
  license = "MIT"
)
}
\arguments{
\item{package_name}{Character string specifying the name of the new package.}

\item{package_path}{Character string specifying the path where the package should be created.
Defaults to the current working directory `"."`.}

\item{github_repo}{Optional character string specifying the GitHub repository in the format
`"username/repo"`. If provided, the function will add it as the remote and push the initial commit.}

\item{author_name}{Character string with the author name for DESCRIPTION and LICENSE.}

\item{author_email}{Character string with the author email for DESCRIPTION.}

\item{license}{Character string specifying the license type. Defaults to `"MIT"`.}
}
\value{
The full path of the newly created package (invisibly).
}
\description{
`af_create_package()` creates a new R package skeleton, initializes Git, optionally sets a GitHub
remote, adds LICENSE and DESCRIPTION metadata, and generates a release runner script for the package.
The release runner script automatically sources `af_packages.R` from the installed `afcommon` library.
}
\details{
Steps performed by this function:
1. Creates the package skeleton using `usethis::create_package()`.
2. Sets DESCRIPTION fields: author, email, license.
3. Adds LICENSE file.
4. Initializes Git and creates the initial commit.
5. Adds GitHub remote and pushes initial commit if `github_repo` is provided.
6. Creates a `release_<package>.R` script in the package root that sources `af_packages.R`
   from the installed `afcommon` library and calls `af_release_package()` for this package.
}
\examples{
\dontrun{
af_create_package(
  package_name = "mypackage",
  package_path = "C:/Users/USER/Dropbox/Amir/Amir personal/R",
  github_repo = "https://github.com/amirhome61/mypackage",
  author_name = "Amir Freund",
  author_email = "amir@email.com"
)
}
}
