% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_factor_analysis.R
\name{af_create_sem_plot}
\alias{af_create_sem_plot}
\title{Create SEM Plot}
\usage{
af_create_sem_plot(
  fit,
  exclude_nodes = NULL,
  layout = "tree",
  nrotation = 1,
  edge_labels = TRUE
)
}
\arguments{
\item{fit}{Fitted lavaan model object}

\item{exclude_nodes}{(Optional) Vector of node names to exclude from the plot}

\item{layout}{Layout style for the plot (default: "tree")}

\item{nrotation}{Rotation of the plot (default: 1)}

\item{edge_labels}{Logical indicating whether to display edge labels (default: TRUE)}
}
\value{
SEM plot object
}
\description{
Generates a structural equation model (SEM) plot using the semPlot package,
with options to exclude specific nodes and customize layout and edge labels.
}
