% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_descriptive.R
\name{af_descriptive}
\alias{af_descriptive}
\title{Comprehensive Descriptive Statistics Function}
\usage{
af_descriptive(df, var_list = NULL)
}
\arguments{
\item{df}{A dataframe to analyze}

\item{var_list}{Optional list of variable names to analyze (default is all variables)}
}
\value{
A list of three GT tables or NULLs for numeric, binary, and categorical variables
}
\description{
Comprehensive Descriptive Statistics Function
}
