% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_utils.R
\name{af_distance}
\alias{af_distance}
\title{calulate the sqrt((A1^2+A2^2 ... An^2)n) for each row}
\usage{
af_distance(...)
}
\arguments{
\item{...}{Numeric vectors representing different dimensions}
}
\value{
Numeric vector of RMS distances for each row
}
\description{
This function calculates the root mean square (RMS) distance for each row
across multiple numeric vectors. It computes the square root of the average
of the squared values for each row, effectively measuring the distance from
the origin in a multi-dimensional space.
}
\examples{
df <- data.frame(A = c(1, 2, 3), B = c(4, 5, 6), C = c(7, 8, 9))
df$new <- af_distance(df$A, df$B, df$C)

}
