% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_factor_analysis.R
\name{af_interpret_bartlett}
\alias{af_interpret_bartlett}
\title{Interpret Bartlett's Test Result}
\usage{
af_interpret_bartlett(bartlett_result)
}
\arguments{
\item{bartlett_result}{Result object from bartlett.test()}
}
\value{
Interpretation string based on the p-value of Bartlett's test
}
\description{
Generates an interpretation of Bartlett's test result for factor analysis suitability.
}
