% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_factor_analysis.R
\name{af_interpret_kmo}
\alias{af_interpret_kmo}
\title{Factor analysis functions}
\usage{
af_interpret_kmo(msa)
}
\arguments{
\item{msa}{Kaiser-Meyer-Olkin measure of sampling adequacy}
}
\value{
Interpretation string for the KMO value
}
\description{
Functions for performing factor analysis, confirmatory factor analysis (CFA),
and measurement invariance testing using the lavaan package.
}
