% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_utils.R
\name{af_is_col_nominal}
\alias{af_is_col_nominal}
\title{Check if Data Frame Column is Unordered Factor with Multiple Levels}
\usage{
af_is_col_nominal(df, col_name)
}
\arguments{
\item{df}{(data.frame) The data frame containing the column}

\item{col_name}{(character) Name of the column to test}
}
\value{
(logical) TRUE if the column is an unordered factor with more than two levels, FALSE otherwise
}
\description{
Tests whether a specified column in a data frame is an unordered (nominal) factor with
more than two levels. Identifies nominal categorical variables for appropriate statistical
treatment in analyses.
}
