% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_utils.R
\name{af_is_col_ordered}
\alias{af_is_col_ordered}
\title{Check if Data Frame Column is Ordered Factor}
\usage{
af_is_col_ordered(df, col_name)
}
\arguments{
\item{df}{(data.frame) The data frame containing the column}

\item{col_name}{(character) Name of the column to test}
}
\value{
(logical) TRUE if the column is an ordered factor, FALSE otherwise
}
\description{
Tests whether a specified column in a data frame is an ordered factor, regardless of
the number of levels. Returns TRUE for any ordered factor variable, useful for determining
if ordinal methods should be applied.
}
