% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_utils.R
\name{af_is_col_surv}
\alias{af_is_col_surv}
\title{Check if Data Frame Column is Survival Object}
\usage{
af_is_col_surv(df, col_name)
}
\arguments{
\item{df}{(data.frame) The data frame containing the column}

\item{col_name}{(character) Name of the column to test}
}
\value{
(logical) TRUE if the column is a Surv object, FALSE otherwise
}
\description{
Tests whether a specified column in a data frame is a Surv object from the survival
package. Used to identify survival time variables appropriate for survival analysis
methods like Cox regression or Kaplan-Meier estimation.
}
