% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_utils.R
\name{af_is_numeric}
\alias{af_is_numeric}
\title{Check if Vector is Numeric}
\usage{
af_is_numeric(v)
}
\arguments{
\item{v}{(vector) The vector to test}
}
\value{
(logical) TRUE if the vector is numeric, FALSE otherwise
}
\description{
Tests whether a vector is numeric. A simple wrapper around is.numeric for consistency
with other type-checking functions in this package.
}
