% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_utils.R
\name{af_last}
\alias{af_last}
\title{Get Last Element of Vector}
\usage{
af_last(v)
}
\arguments{
\item{v}{(vector) A vector from which to extract the last element}
}
\value{
The last element of the vector, maintaining its original type
}
\description{
Returns the last element of a vector. A convenience function that extracts the final
value, useful for obtaining the most recent observation, maximum level, or endpoint
in ordered sequences.
}
