% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_utils.R
\name{af_order}
\alias{af_order}
\title{Sort Vector in Ascending Order}
\usage{
af_order(l)
}
\arguments{
\item{l}{(vector) A vector to be sorted}
}
\value{
(vector) The input vector sorted in ascending order
}
\description{
Returns a vector sorted in ascending order. A simple wrapper around the order function
that sorts elements from smallest to largest. Works with numeric, character, and other
orderable data types.
}
