% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_basic_plots.R
\name{af_plot_barchart}
\alias{af_plot_barchart}
\title{Plot Bar Chart}
\usage{
af_plot_barchart(df, y_name)
}
\arguments{
\item{df}{(data.frame) The data frame containing the variable to plot}

\item{y_name}{(character) The name of the column to plot}
}
\value{
(ggplot) A ggplot object displaying the bar chart with count on the y-axis
}
\description{
Creates a bar chart for a categorical or discrete variable in a data frame. The plot
displays count frequencies with blue fill and includes an annotation showing the count
of non-missing values in the upper right corner.
}
\examples{
af_plot_barchart(mtcars, "cyl")

}
