% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_basic_plots.R
\name{af_plot_correlation_per_group}
\alias{af_plot_correlation_per_group}
\title{Plot Correlation Matrices by Group}
\usage{
af_plot_correlation_per_group(df, group_var, group_values = NULL, vars = NULL)
}
\arguments{
\item{df}{(data.frame) The data frame containing the variables to correlate and the grouping variable}

\item{group_var}{(character) The name of the grouping variable}

\item{group_values}{(vector) The specific group values to create plots for. Default is NULL, which uses all unique values of the grouping variable}

\item{vars}{(character vector) The names of variables to include in the correlation plots. Default is NULL, which uses all variables in the data frame}
}
\value{
(list) A list of ggplot objects, one correlation plot for each group value
}
\description{
Creates separate correlation plots for each level of a grouping variable. For each
group value, the function filters the data and generates a correlation matrix plot
with the group identifier in the title. Returns a list of correlation plots.
}
\examples{
plots <- af_plot_correlation_per_group(mtcars, "cyl")
plots <- af_plot_correlation_per_group(mtcars, "cyl", c(4, 6), c("mpg", "hp", "wt"))

}
