% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_basic_plots.R
\name{af_plot_distribution}
\alias{af_plot_distribution}
\title{Plot Distribution with Histogram and Density Curve}
\usage{
af_plot_distribution(df, y_name, title = " ", nbins = 50)
}
\arguments{
\item{df}{(data.frame) The data frame containing the variable to plot}

\item{y_name}{(character) The name of the column to plot}

\item{title}{(character) The plot title. Default is " "}

\item{nbins}{(numeric) The number of bins for the histogram when plotting numeric variables. Default is 50}
}
\value{
(ggplot) A ggplot object displaying the distribution
}
\description{
Creates a distribution plot for a variable in a data frame. For numeric
variables, generates a histogram overlaid with a kernel density curve and
displays total sample size. For non-numeric variables, creates a bar plot.
}
\examples{
# Numeric variable with density overlay
af_plot_distribution(mtcars, "mpg", title = "MPG Distribution")

}
