% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_basic_plots.R
\name{af_plot_distribution_per_group}
\alias{af_plot_distribution_per_group}
\title{Plot Distributions by Group}
\usage{
af_plot_distribution_per_group(df, y_name, group_var, group_values = NULL)
}
\arguments{
\item{df}{(data.frame) The data frame containing the variable to plot and the grouping variable}

\item{y_name}{(character) The name of the column to plot}

\item{group_var}{(character) The name of the grouping variable}

\item{group_values}{(vector) The specific group values to create plots for. Default is NULL, which uses all unique values of the grouping variable}
}
\value{
(list) A list of ggplot objects, one distribution plot for each group value
}
\description{
Creates separate distribution plots for a variable across different levels of a grouping
variable. For each group value, the function filters the data and generates a distribution
plot with the group identifier in the title. Returns a list of distribution plots.
}
\examples{
plots <- af_plot_distribution_per_group(mtcars, "mpg", "cyl")
plots <- af_plot_distribution_per_group(mtcars, "mpg", "cyl", c(4, 6))

}
