% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_utils.R
\name{af_remove_empty_vars}
\alias{af_remove_empty_vars}
\title{Remove variables from var_list that are completely empty in df}
\usage{
af_remove_empty_vars(df, var_list)
}
\arguments{
\item{df}{Data frame to check for empty variables}

\item{var_list}{Character vector of variable names to check}
}
\value{
Character vector of variable names from var_list that have non-missing data in df
}
\description{
This function checks a data frame for variables listed in var_list and removes
those that contain only missing values (NA or NULL). It returns a character vector
of variable names that have at least some non-missing data.
}
\examples{
df <- data.frame(A = c(1, 2, NA), B = c(NA, NA, NA), C = c(3, 4, 5))
var_list <- c("A", "B", "C")
af_remove_empty_vars(df, var_list)


}
