% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_utils.R
\name{af_reverse_scale}
\alias{af_reverse_scale}
\title{reverse a survey question scale (from 1..7 to 7..1)}
\usage{
af_reverse_scale(score, scale_min, scale_max)
}
\arguments{
\item{score}{Numeric vector of survey scores}

\item{scale_min}{Numeric value representing the minimum value of the scale}

\item{scale_max}{Numeric value representing the maximum value of the scale}
}
\value{
Numeric vector of reversed scores
}
\description{
This function reverses a survey question scale by transforming each score
based on the provided minimum and maximum scale values.
It effectively flips the scale, so that higher scores become lower scores and vice versa.
}
\examples{
scores <- c(1, 2, 3, 4, 5, 6, 7)
af_reverse_scale(scores, scale_min = 1, scale_max = 7)

}
