% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_utils.R
\name{af_select_val}
\alias{af_select_val}
\title{Conditional Value Selection}
\usage{
af_select_val(cond, a, b)
}
\arguments{
\item{cond}{(logical) The condition to evaluate}

\item{a}{The value to return if condition is TRUE}

\item{b}{The value to return if condition is FALSE}
}
\value{
Either a or b depending on the condition
}
\description{
Returns one of two values based on a logical condition. A simple if-else wrapper that
returns the first value if the condition is TRUE, otherwise returns the second value.
Useful for inline conditional assignments.
}
